Move another 20 classes out of internal
