+++ core/src/org/labkey/core/script/ScriptableErrorsList.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.labkey.api.script;
+package org.labkey.core.script;
 
 import org.labkey.api.query.BatchValidationException;
 import org.labkey.api.query.ValidationException;
