+++ pipeline/src/org/labkey/pipeline/api/ExecTaskFactory.java
@@ -43,7 +43,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import static org.labkey.api.util.CommandLineTokenizer.tokenize;
+import static org.labkey.pipeline.api.CommandLineTokenizer.tokenize;
 
 /**
  * User: kevink
